zephyr: Fix format specifiers in custom timestamp formatter
Conversion specifier %d is not valid for uint64_t type. To achieve
portable format we should use format macro constants e.g. PRIu32 for
uint32_t.

BUG=b:320259967
TEST=Build firmware with 'zmake build bloonchipper'. Make sure timestamp
     is correctly printed in Zephyr log lines.

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5273503
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>

+++ zephyr/shim/src/console.c
@@ -512,8 +512,9 @@ static int custom_timestamp(const struct log_output *output,
 {
 	uint64_t us = log_output_timestamp_to_us(timestamp);
 
-	return printer(output, "[%d.%06d] ", us / USEC_PER_SEC,
-		       us % USEC_PER_SEC);
+	return printer(output, "[%" PRIu32 ".%06" PRIu32 "] ",
+		       (uint32_t)(us / USEC_PER_SEC),
+		       (uint32_t)(us % USEC_PER_SEC));
 }
 
 static int timestamp_init(void)
